+++ src/MarketplaceMetering/MarketplaceMeteringClient.php
@@ -1,0 +1,11 @@
+<?php
+namespace Aws\MarketplaceMetering;
+
+use Aws\AwsClient;
+
+/**
+ * This client is used to interact with the **AWSMarketplace Metering** service.
+ * @method \Aws\Result meterUsage(array $args = [])
+ * @method \GuzzleHttp\Promise\Promise meterUsageAsync(array $args = [])
+ */
+class MarketplaceMeteringClient extends AwsClient {}